+++ CHANGELOG.md
@@ -3,8 +3,8 @@
 ## [1.3.1] - 2025-08-18
 
 ### Added
+- Match system theme toggle
 - Open in window link
-- Added drag-and-drop/hide navigation customization
 
 ## [1.3.0] - 2025-08-13